LayoutTests/imported/w3c/web-platform-tests/css/css-scroll-snap/snap-after-relayout/adding-only-snap-area-expected.txt


FAIL Adding a new snap area when there are none should make the scroller snap to it. assert_equals: expected 100 but got 0
